+++ R/char_select.R
@@ -51,7 +51,8 @@ char_select.character <- function(x, pattern, selection = c("keep", "remove"),
     ids <- object2id(pattern, types = x, 
                      valuetype = valuetype, 
                      case_insensitive = case_insensitive,
-                     concatenator = " ")
+                     concatenator = " ",
+                     match_pattern = "single")
     id <- unlist_integer(ids)
     
     if (selection == "keep") {