+++ NEWS.md
@@ -6,6 +6,9 @@
 
 * Further improve the verbose messages for corpus, tokens, dfm and fcm objects.
 
+* Changes to dfm_select() now include new arguments `match_pattern` and `concatenator`, to control how single-word patterns or multi-word patterns
+  and multi-word patterns matched, as well as controlling the concatenator that binds phrases in the pattern.
+
 # quanteda 4.2.0
 
 ## Changes and additions