+++ NEWS.md
@@ -6,6 +6,15 @@
 
 * Significantly reduce the memory usage for the `c` operation on large `tokens` and `tokens_xptr` objects.
 
+* Further improvements to the verbose messages for corpus, tokens, dfm and fcm objects.
+
+* `tokens_ngrams()` now includes a new argument `apply_if`, functioning similar to this argument
+in `tokens_compound()` and `tokens_lookup()` (#2390).
+
+
+* Changes to dfm_select() now include new arguments `match_pattern` and `concatenator`, to control how single-word patterns or multi-word patterns
+  and multi-word patterns matched, as well as controlling the concatenator that binds phrases in the pattern.
+
 # quanteda 4.2.0
 
 ## Changes and additions