+++ NEWS.md
@@ -15,6 +15,8 @@ in `tokens_compound()` and `tokens_lookup()` (#2390).
 * Changes to dfm_select() now include new arguments `match_pattern` and `concatenator`, to control how single-word patterns or multi-word patterns
   and multi-word patterns matched, as well as controlling the concatenator that binds phrases in the pattern.
 
+* `data_corpus_inaugural` now updated for Trump 2025.
+
 # quanteda 4.2.0
 
 ## Changes and additions